Refactor nodule segmentation and consensus mask generation
This commit refactors the Nodule class in the dataset module to improve the segmentation of nodules and the generation of consensus masks. The `segment_nodule` method has been renamed to `get_consensus_mask` and now returns the consensus mask of the nodule based on the annotations of the 4 radiologists. The `segment_nodule` method has been updated to use the `get_consensus_mask` method and now cuts out the nodule from the scan using the consensus boundary or removes the nodule from the scan based on the `invert` parameter.